requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv